requirements.txt
python-dotenv==0.13.0
python-telegram-bot==12.7
requests==2.24.0
selenium==3.141.0